Fixes button for fetching data
The XMLHttpRequest was never actually being sent. Adding
request.send() to the end of the function makes the actual request
happen. The second bug was status === 20. HTTP statuses start at
100 and go up to 500. Anything in the 200 range is a successful
response with 200 being the default success response status.

The code written at this point prints the JSON object to the console.

References:
https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/send
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status